+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/SampleProvider.java
@@ -5,6 +5,7 @@ public interface SampleProvider {
     byte PROVIDER_PEBBLE_MORPHEUZ = 1;
     byte PROVIDER_PEBBLE_GADGETBRIDGE = 2;
     byte PROVIDER_PEBBLE_MISFIT = 3;
+    byte PROVIDER_PEBBLE_HEALTH = 4;
 
     byte PROVIDER_UNKNOWN = 100;