+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/pebble/PebbleCoordinator.java
@@ -49,7 +49,7 @@ public class PebbleCoordinator extends AbstractDeviceCoordinator {
         SharedPreferences sharedPrefs = PreferenceManager.getDefaultSharedPreferences(GBApplication.getContext());
         if (sharedPrefs.getBoolean("pebble_force_untested", false)) {
             //return new PebbleGadgetBridgeSampleProvider();
-            return new MisfitSampleProvider();
+            return new HealthSampleProvider();
         } else {
             return new MorpheuzSampleProvider();
         }